Fix "apply: wrong number of arguments (3 . 3), 0"
A warning was generated on startup.
The root problem was in a call to (color-rgb-to-hsl) which expects three
arguments but got none in this call stack:
  (color-rgb-to-hsl)
  (apply color-rgb-to-hsl nil)
  (color-lighten-name "unspecified-bg" -3)
  (color-darken-name "unspecified-bg" 3)
  [[file:~/Dropbox/sync-linux/.emacs.d/personal/personal.el:::background%20(color-darken-name][(set-face-attribute 'auto-dim-other-buffers-face nil :background (color-darken-name (face-attribute 'default :background) 3))]]
  (adjust-dim-face solarized-dark t)
  (apply adjust-dim-face (solarized-dark t))
  (load-theme solarized-dark t)
  (solarized)
  (if (tty-type (frame-terminal)) (zenburn) (solarized))
  (progn (advice-add 'load-theme :after #'adjust-dim-face) (if (tty-type (frame-terminal)) (zenburn) (solarized)))
  ((lambda nil (progn (advice-add 'load-theme :after #'adjust-dim-face) (if (tty-type (frame-terminal)) (zenburn) (solarized)))))
  (run-hooks emacs-startup-hook term-setup-hook)

The solution is to check for unspecified-bg color, which seems to
occur when the terminal is headless, as in a daemon.

Also, remove extraneous insinuation of theme when processing
jeff/organizer.  The theme should have been set via the emacs-startup-hook.